+++ includes/features/class-analytics.php
@@ -2070,7 +2070,7 @@ class Analytics {
 		foreach ( $replace as $key => $arg ) {
 			$params[ $key ] = $arg;
 		}
-		$url = admin_url( 'tools.php?page=traffic-viewer' );
+		$url = admin_url( 'admin.php?page=traffic-viewer' );
 		foreach ( $params as $key => $arg ) {
 			if ( '' !== $arg ) {
 				$url .= '&' . $key . '=' . $arg;